first_impressions_mapping/requirements.txt
cmake
termcolor
numpy
opencv-python
scikit-image
python-interface
scikit-learn
dlib
